Rename Entities namespace to Database, and move DbContextFactory to the Database namespace
